requirements.txt
Django==1.7
argparse==1.2.1
django-ckeditor==4.4.6
django-debug-toolbar==1.2.1
pytz==2014.7
sqlparse==0.1.13
wsgiref==0.1.2
